pyproject.toml
[tool.poetry]
name = "recoservice"
version = "0.1.0"
description = "Service for Recomedation System course"
authors = ["Danis <bzakks21@gmail.com>"]
license = "Apache License"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.1"
uvicorn = { extras = ["standard"], version = "^0.24.0.post1" }
orjson = "^3.9.10"
pydantic-settings = "^2.0.3"
pyjwt = "^2.8.0"
httpx = "^0.25.1"
gunicorn = "^21.2.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"
nest-asyncio = "^1.5.8"
pytest-mock = "^3.12.0"


[tool.poetry.group.lint.dependencies]
ruff = "^0.1.5"
black = "^23.11.0"
isort = "^5.12.0"
pylint = "^3.0.2"
flake8 = "^6.1.0"
mypy = "^1.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
target-version = ["py311"]

[tool.ruff]
fix = true
line-length = 120
respect-gitignore = true

[tool.ruff.isort]
case-sensitive = true


[tool.pytest]
console_output_style = "progress"
testpaths = "tests"
junit_family = "xunit2"

[tool.pytest.ini_options]
asyncio_mode = 'auto'
